Introduce java.lang.invoke.VarHandle APIs.
These are present in preparation for OpenJDK 9 to allow VM support for
VarHandle accessors and fences to be developed. For this initial
import the VarHandle APIs are hidden and unimplemented.

The implementation path will follow the approach taken with the
MethodHandle API: public API compatability and alternative
implementation strategy based on VM functionality.

The files in this CL under ojluni/ define the APIs and are based on
the corresponding files in upstream OpenJDK 9+181.

Bug: 65872996
Test: m -j32
BenzoEdit: Fixed up for master->oreo
Change-Id: I678ab2239507b349a9701a4a6f5bbc92bf18b40d
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>