Libcore: Clean up library loading
Remove Runtime.doLoad, remove library path parameter to nativeLoad.
Move the setting to be retrieved by ART, to hide the complexity
and make similar library loading easier for future changes.

This also removes an old and somewhat stale comment for the need of
passing the LD_LIBRARY_PATH, and a no longer necessary synchronization
block.

Bug: 70901841
Test: m test-art-host
Test: device boots
Change-Id: Iaccaeb56422877abac9f7fe6f5a17364c8adf4ca
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>