Restore the visibility checks in Atomic.*FieldUpdater
This change is just reverting unintended diffs against upstream
sources. (The call to checkPackageAccess is a noop on Android because
it only does SecurityManager checks, so this change leaves that
suppresses as it is right now. The call to ensureMemberAccess does
regular visibility checks, and is restored in this change.)

Bug: 75269521
Test: cts-tradefed run cts-dev -m CtsLibcoreJsr166TestCases
Test: cts-tradefed run cts-dev -m CtsLibcoreTestCases
Change-Id: I5e050f786709fb28543ca4ea409eb9876539fc6f
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>