Restore more visibility checks in AtomicLongFieldUpdater
This change was accidentally omitted from aosp/644539. That change
restored the check when AtomicLong.VM_SUPPORTS_LONG_CAS is true
only. This change restores it when that value is false.

Before this change, AtomicIntegerFieldUpdaterTest in jsr166-tests was
failing in an environment where that value is false. This change must
fixes that.

Bug: 75269521
Test: `cts-tradefed run cts-dev -m CtsLibcoreJsr166TestCases -t jsr166.AtomicLongFieldUpdaterTest` in a build where AtomicLong.VM_SUPPORTS_LONG_CAS is hardcoded to false
Change-Id: I155f427ba624e4ea293a0983de4f7e65c24aff55
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>